refactor: remove dispatch methods from Host trait
Host no longer has dispatch_ask_user, dispatch_delegate, or
dispatch_mcp. Handler factories capture their dependencies directly:

- ask_user captures pending_asks (shared Arc with NodeHost for
  reply_to_ask delivery)
- delegate captures event_tx (NodeEventSender)
- mcp captures Arc<McpHandler>

Host is now purely lifecycle and session state — no tool dispatch.
