+++ crates/node/src/tools/ask_user.rs
@@ -1,13 +1,16 @@
 //! Ask-user tool handler factory.
 
-use runtime::host::Host;
 use serde::Deserialize;
-use std::sync::Arc;
+use std::{collections::HashMap, sync::Arc, time::Duration};
+use tokio::sync::{Mutex, oneshot};
 use wcore::{
     ToolDispatch, ToolHandler,
     agent::{AsTool, ToolDescription},
     model::Tool,
 };
+
+/// Timeout for waiting on user reply (5 minutes).
+const ASK_USER_TIMEOUT: Duration = Duration::from_secs(300);
 
 /// A single option the user can choose from.
 #[derive(Deserialize, schemars::JsonSchema)]
@@ -43,14 +46,42 @@ impl ToolDescription for AskUser {
     const DESCRIPTION: &'static str = r#"Ask the user one or more structured questions with predefined options. Each question needs a short UI header, the full question text, and options with labels and descriptions. The user picks from the options or types a free-text "Other" answer. Returns JSON mapping question text to selected label. For multi_select, the answer is a comma-joined string like "Option A, Option B"."#;
 }
 
-pub fn handler<H: Host + 'static>(host: H) -> (Tool, ToolHandler) {
+/// Pending ask_user oneshots, keyed by conversation_id.
+pub type PendingAsks = Arc<Mutex<HashMap<u64, oneshot::Sender<String>>>>;
+
+pub fn handler(pending_asks: PendingAsks) -> (Tool, ToolHandler) {
     (
         AskUser::as_tool(),
         Arc::new(move |call: ToolDispatch| {
-            let host = host.clone();
+            let pending_asks = pending_asks.clone();
             Box::pin(async move {
-                host.dispatch_ask_user(&call.args, call.conversation_id)
-                    .await
+                let input: AskUser = serde_json::from_str(&call.args)
+                    .map_err(|e| format!("invalid arguments: {e}"))?;
+
+                let conversation_id = call
+                    .conversation_id
+                    .ok_or("ask_user is only available in streaming mode")?;
+
+                let (tx, rx) = oneshot::channel();
+                pending_asks.lock().await.insert(conversation_id, tx);
+
+                match tokio::time::timeout(ASK_USER_TIMEOUT, rx).await {
+                    Ok(Ok(reply)) => Ok(reply),
+                    Ok(Err(_)) => {
+                        pending_asks.lock().await.remove(&conversation_id);
+                        Err("ask_user cancelled: reply channel closed".to_owned())
+                    }
+                    Err(_) => {
+                        pending_asks.lock().await.remove(&conversation_id);
+                        let headers: Vec<&str> =
+                            input.questions.iter().map(|q| q.header.as_str()).collect();
+                        Err(format!(
+                            "ask_user timed out after {}s: no reply received for: {}",
+                            ASK_USER_TIMEOUT.as_secs(),
+                            headers.join("; "),
+                        ))
+                    }
+                }
             })
         }),
     )
