+++ crates/node/src/tools/mcp.rs
@@ -1,19 +1,20 @@
 //! MCP tool handler factory.
 
-use runtime::{AgentScope, host::Host};
+use crate::mcp::McpHandler;
+use runtime::AgentScope;
 use std::{
     collections::BTreeMap,
     sync::{Arc, RwLock},
 };
 use wcore::{ToolDispatch, ToolHandler};
 
-/// Build a handler that dispatches MCP tool calls through the host.
-pub fn handler<H: Host + 'static>(
-    host: H,
+/// Build a handler that dispatches MCP tool calls through the McpHandler.
+pub fn handler(
+    mcp: Arc<McpHandler>,
     scopes: Arc<RwLock<BTreeMap<String, AgentScope>>>,
 ) -> ToolHandler {
     Arc::new(move |call: ToolDispatch| {
-        let host = host.clone();
+        let mcp = mcp.clone();
         let scopes = scopes.clone();
         Box::pin(async move {
             let allowed_mcps: Vec<String> = scopes
@@ -23,7 +24,7 @@ pub fn handler<H: Host + 'static>(
                 .filter(|s| !s.mcps.is_empty())
                 .map(|s| s.mcps.clone())
                 .unwrap_or_default();
-            host.dispatch_mcp(&call.args, &allowed_mcps).await
+            crate::mcp::dispatch::dispatch_mcp(&mcp, &call.args, &allowed_mcps).await
         })
     })
 }
